requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
pymongo==4.13.1
python-dotenv==1.1.0
openai==1.86.0
google-generativeai==0.8.5
pydantic==2.7.1